+++ src/typing_ex/typing.cljs
@@ -15,7 +15,8 @@
 (def ^:private todays-limit 10)
 
 ;; bump-version.sh will rewrite
-(def ^:private timeout 60)
+(def ^:private timeout
+  (int (.-value (.getElementById js/document "timeout"))))
 
 (def interval (atom 1000)) ;; milli second